test: don't remove temp dir if debugging

+++ tests/globalTeardown.ts
@@ -1,11 +1,13 @@
 import type { FullConfig } from '@playwright/test'
 import fs from 'fs/promises'
-import { tempDirName } from './utils/index.js'
+import { isDebug, tempDirName } from './utils/index.js'
 
 async function globalTeardown(_config: FullConfig) {
   const dest = new URL(`../${tempDirName}/`, import.meta.url)
 
-  await fs.rm(dest, { force: true, recursive: true })
+  if (!isDebug) {
+    await fs.rm(dest, { force: true, recursive: true })
+  }
 }
 
 export default globalTeardown
